+++ src/main/java/pl/asie/computronics/integration/railcraft/block/BlockTicketMachine.java
@@ -25,7 +25,6 @@ public class BlockTicketMachine extends BlockMachineSidedIcon implements IBlockW
 		super("ticket_machine");
 		this.setBlockName("computronics.ticketMachine");
 		this.setGuiProvider(Computronics.railcraft.guiTicketMachine);
-		this.setNoNedoComputers(true);
 	}
 
 	@Override