+++ src/main/java/pl/asie/computronics/tile/TileCipherBlockAdvanced.java
@@ -299,22 +299,4 @@ public class TileCipherBlockAdvanced extends TileEntityPeripheralBase {
 		}
 		return null;
 	}
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public boolean connectable(int side) {
-		return false;
-	}
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public short busRead(int addr) {
-		return 0;
-	}
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public void busWrite(int addr, short data) {
-
-	}
 }